fix header location, add permission

+++ handlers/element/delete.php
@@ -2,16 +2,21 @@
 
 session_start();
 if(!isset($_SESSION["name"])){
-    header('Location: login/index.php');
+    header('Location: ../../login/index.php');
+}else{
+    if(!$_SESSION['edit']){
+        header('Location: ../../index.php?board='.$_SESSION['board']);
+    }else{
+
+        include '../../src/cfg/conn.php';
+
+        $sql = "DELETE FROM `element` WHERE `element`.`id_element` = ".$_GET['id'];
+
+        $res = @mysqli_query($conn, $sql);
+
+        mysqli_close($conn);
+
+        header('Location: ../../index.php?board='.$_SESSION['board']);
+    }
 }
-
-include '../../src/cfg/conn.php';
-
-$sql = "DELETE FROM `element` WHERE `element`.`id_element` = ".$_GET['id'];
-
-$res = @mysqli_query($conn, $sql);
-
-mysqli_close($conn);
-
-header('Location: ../../index.php?board='.$_SESSION['board']);
 ?>
